app/requirements.txt
django
djangorestframework
psycopg2-binary==2.8.5
